Added little spaceships

+++ GamePanel.java
@@ -6,33 +6,48 @@ import java.awt.Graphics2D;
 import java.awt.geom.Area;
 
 import javax.swing.JPanel;
+import javax.swing.border.TitledBorder;
 
 
 /**
  * GamePanel
  */
+
+enum GameState {
+  TitleScreen,
+  GameScreen,
+  DeathScreen,
+  ExitScreen,
+  PrepareToRun
+};
+
 public class GamePanel extends JPanel implements Runnable{
   // You won't win over my mind sun!
-  final int originalTileSize = 16;
-  final int scale = 3;
+  final static int originalTileSize = 16;
+  final static int scale = 3;
 
-  final int tileSize = originalTileSize * scale;
-  final int maxScreenCol = 16;
-  final int maxScreenRow = 12;
-  final int screenWidth = tileSize*maxScreenCol;
-  final int screenHeight = tileSize*maxScreenRow;
+  final static int tileSize = originalTileSize * scale;
+  final static int maxScreenCol = 20;
+  final static int maxScreenRow = 12;
+  final static int screenWidth = tileSize*maxScreenCol;
+  final static int screenHeight = tileSize*maxScreenRow;
 
   int time = 0;
   int FPS = 144;
-  boolean pageSettings = false;
+  GameState gameState;
 
+  // System
   Thread gameThread;
   KeyHandler keyHandler = new KeyHandler();
+  Sound sound = new Sound();
+  GUI gui = new GUI(screenWidth, screenHeight);
 
-  PlayerHandler playerH = new PlayerHandler(screenWidth, screenHeight, 0.2f, 10);
+  // Objects
+  PlayerHandler playerH = new PlayerHandler(screenWidth, screenHeight, 0.15f, 10);
   AsteroidsHandler asteroidsHandler = new AsteroidsHandler(screenWidth, screenHeight);
   ParticleSystem particleSystem = new ParticleSystem();
-  GUI gui = new GUI(screenWidth, screenHeight);
+  AlienHandler alienHandler = new AlienHandler(5);
+
 
   public GamePanel(){
     this.setPreferredSize(new Dimension(screenWidth, screenHeight));
@@ -40,6 +55,10 @@ public class GamePanel extends JPanel implements Runnable{
     this.setDoubleBuffered(true);
     this.addKeyListener(keyHandler);
     this.setFocusable(true);
+
+    alienHandler.add();
+
+    gameState = GameState.TitleScreen;
   }
   public void startGameThread(){
     gameThread = new Thread(this);
@@ -54,11 +73,11 @@ public class GamePanel extends JPanel implements Runnable{
     long timer = 0;
     int drawCount = 0;
     time++;
+    sound.setFile(0);
 
     while (gameThread != null) {
       currentTime = System.nanoTime();
       time++;
-
       delta += (currentTime - lastTime) / drawInterval;
       timer += (currentTime - lastTime);
       lastTime = currentTime;
@@ -69,18 +88,29 @@ public class GamePanel extends JPanel implements Runnable{
         drawCount++;
       }
       if ( timer >= 1000000000){
-        //System.out.println("FPS: " + drawCount);
+        System.out.println("FPS: " + drawCount);
         drawCount = 0;
         timer = 0;
       }
-
+      if (gameState == GameState.ExitScreen) System.exit(0);
     }
   }
   public void update(){
+    if (gameState == GameState.TitleScreen )return;
+    if (gameState == GameState.PrepareToRun){
+      gameState = GameState.GameScreen;
+      playerH.reset();
+      asteroidsHandler.reset();
+      playerH.ballotsHandler.reset();
+      gui.points = 0;
+      return;
+    }
     playerH.React(keyHandler.leftPressed, keyHandler.rightPressed, keyHandler.upPressed, keyHandler.spacePressed);
     playerH.ballotsHandler.update();
     asteroidsHandler.update();
     particleSystem.update();
+    alienHandler.update(asteroidsHandler.asteroids, playerH);
+
     if (Math.random() < 0.005 && asteroidsHandler.active_asteroids() < asteroidsHandler.MAX_ASTEROIDS/2.0){
       time++;
       asteroidsHandler.add_asteroid(Math.max((float)Math.random()*60, 30.0f));
@@ -88,19 +118,25 @@ public class GamePanel extends JPanel implements Runnable{
     // Collision detection
     float collision;
     for (Asteroid asteroid: asteroidsHandler.asteroids) {
+      if (asteroid == null) break;
       if (asteroid.active){
         collision = CollisionHandler.collisionSAT(asteroid.shape, playerH.shape, asteroid, playerH);
         if (collision != 10000f){
           particleSystem.add_boom_particles(playerH, 100f);
           playerH.kill();
+          gameState = GameState.DeathScreen;
+          //sound.setFile(0);
+          sound.play();
+          //playSE(0);
         }
       }
     }
     for (Asteroid asteroid: asteroidsHandler.asteroids) {
+      if (asteroid == null) break;
       if (asteroid.active && playerH.ballotsHandler.isColliding(asteroid)){
-        if (asteroid.size_shape > 20.0f){
-          System.out.println(asteroidsHandler.splitAsteroid(asteroid));
-        }
+        if (asteroid.size_shape > 20.0f)
+          asteroidsHandler.splitAsteroid(asteroid);
+        sound.play();
         gui.addPoints(asteroid.size_shape);
         particleSystem.add_boom_particles(asteroid, asteroid.size_shape);
         asteroid.kill();
@@ -115,16 +151,33 @@ public class GamePanel extends JPanel implements Runnable{
 
     g2.setFont(new Font("serif", Font.PLAIN, 32));
     g2.setColor(Color.white);
-
-
-    playerH.Draw(g2);
-    playerH.DrawBullets(g2);
-    asteroidsHandler.Draw(g2);
-    particleSystem.Draw(g2);
-    gui.Draw(g2, playerH.isAlive);
-
+    if (gameState == GameState.TitleScreen){
+      gameState = gui.DrawTitleScreen(g2, gameState, keyHandler);
+    }
+    if (gameState == GameState.GameScreen || gameState == GameState.DeathScreen){
+      playerH.Draw(g2);
+      playerH.DrawBullets(g2);
+      asteroidsHandler.Draw(g2);
+      particleSystem.Draw(g2);
+      alienHandler.Draw(g2);
+      gui.DrawGameUI(g2, playerH.isAlive);
+      if (gameState == GameState.DeathScreen){
+        gameState = gui.DrawDeathUI(g2, keyHandler);
+      }
+    }
     
     g2.dispose();
 
+  }
+
+  public void playMusic(int i){
+    sound.setFile(i);
+    sound.play();
+    sound.loop();
+  }
+  public void stopMusic(int i){sound.stop();}
+  public void playSE(int i){
+    sound.setFile(i);
+    sound.play();
   }
 }
